+++ unocss/src/rules/colors.ts
@@ -1,31 +1,19 @@
 import { DynamicRule } from "@unocss/core";
 import Theme, { ThemeColor } from "../theme/Theme";
 
-const staticColorCss = (color: string) => `hsl(var(--hue), var(--col-${color}-s), var(--col-${color}-l), var(--col-${color}-a))`;
-const interactiveColorCss = (color: string) => `hsl(var(--hue), var(--col-${color}-s), calc(var(--col-${color}-l) + var(--highlight)), var(--col-${color}-a))`;
-
-const tryCompileValue = (value: string, theme: Theme): string | undefined => {
-  const color = theme.windblade.colors[value];
-  if (color !== undefined) {
-    if (color.interactive) {
-      return interactiveColorCss(value);
-    } else {
-      return staticColorCss(value);
-    }
-  }
-
-  return theme.windblade.miscColors[value];
-}
-
 const colorRule = (prefix: string, property: string): DynamicRule<Theme> => {
   return [
     new RegExp(`^(${prefix})-(.+)$`),
     (match, { theme }) => {
-      const css: any = {};
-      const color = tryCompileValue(match[2], theme);
-      if (!color) return undefined;
-      css[property] = color;
-      return css;
+      if (theme.windblade.miscColors[match[2]]) return {
+        [property]: match[2]
+      }
+
+      if (theme.windblade.colors[match[2]]) return {
+        [property]: `var(--${match[2]}-0)`
+      };
+
+      return;
     }
   ];
 };
@@ -40,23 +28,17 @@ const colorBgRule = (prefix: string): DynamicRule<Theme> => {
   return [
     new RegExp(`^(${prefix})-(.+)$`),
     (match, { theme }) => {
-      const css: any = {};
-      const color = theme.windblade.colors[match[2]];
+      const color = theme.windblade.colors[match[2]]
       if (!color) return;
 
-      if (color.interactive) {
-        css['background-color'] = interactiveColorCss(match[2]);
-        color.on.forEach((_, i) => {
-          css[`--fg-${i + 1}`] = interactiveColorCss(`on-${match[2]}-${i}`);
-        });
-      } else {
-        css['background-color'] = staticColorCss(match[2]);
-        color.on.forEach((_, i) => {
-          css[`--fg-${i + 1}`] = staticColorCss(`on-${match[2]}-${i}`);
-        });
-      }
+      const css: any = {
+        'background': `var(--${match[2]}-0)`,
+        'color': 'var(--fg-1)', // this CSS varaible gets defined below
+      };
 
-      css['color'] = 'var(--fg-1)';
+      for (let i = 1; i <= color.on.length; ++i) {
+        css[`--fg-${i}`] = `var(--${match[2]}-${i})`;
+      }
 
       return css;
     }
@@ -67,9 +49,13 @@ const fgColorRule = (prefix: string, property: string): DynamicRule<Theme> => {
   return [
     new RegExp(`^(${prefix})-(.+)$`),
     (match, { theme }) => {
-      const css: any = {};
-      css[property] = `var(--fg-${match[2]})`
-      return css;
+      if (theme.windblade.miscColors[match[2]]) return {
+        [property]: match[2],
+      }
+
+      return {
+        [property]: `var(--fg-${match[2]})`,
+      };
     }
   ];
 };